Added more styling and grouped components in directories
