+++ src/main/Loader.js
@@ -1,0 +1,13 @@
+import React from "react";
+function Loader() {
+  return (
+    <div className="loader">
+      <div class="lds-ripple">
+        <div></div>
+        <div></div>
+      </div>
+    </div>
+  );
+}
+
+export default Loader
